Fixed server crashing error #1, minor logging updates
Fixed a bug where the server crashes when attempting to kill a disconnected player by checking if the player exists before attempting to kill them. Instead, if the player no longer exists the server will log the attempt.

This now shows that there is more work to be done to fix this bug, as the server will attempt to kill the non-existent player indefinitely.

I also updated the server logging slightly to show the player's chosen name on connect.